Phase 0-1: modernize project layout and build for 64-bit
- Add README documenting the original tech stack and known blockers
- Remove xcuserdata, pbxuser and SCons build caches; add .gitignore
- Add a shared Grid scheme (the originals lived under xcuserdata)
- Drop the dangling iTunesArtwork resource reference
- Switch ARCHS to $(ARCHS_STANDARD)
- Guard kazmath's 32-bit NEON assembly behind KM_USE_NEON_ASM so arm64
  falls back to the pure C matrix multiply
- Make __ccContentScaleFactor CGFloat consistently across headers
- Cast target_ to CCNode* in CCRotateTo/CCRotateBy so -rotation resolves
- Remove iAd (the banner calls were already commented out) and the
  unreferenced CCVideoPlayer extension
- Drop UIAccelerometer support from CCLayer; the game never used it
- Fix 64-bit truncation, float abs and format-string issues in app code

+++ Grid/GCHelper.m
@@ -449,7 +449,7 @@ static GCHelper *sharedHelper = nil;
             
             // process the score information.
             
-               int leaderboardscore=leaderboardRequest.localPlayerScore.value;
+               int leaderboardscore=(int)leaderboardRequest.localPlayerScore.value;
              NSLog(@"old score%d",leaderboardscore);
             if([GCState sharedInstance].pvpSocre<leaderboardscore)
             {
